+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 * Better specs
 * Unimplemented i18next methods raise an exception
 * Support methods exists, languages, loadNamespaces and resolvedLanguage
-
 * Add .yardopts to the gem file to show documentation at rubydoc.info
+* Specs use a local copy of the i18next-fetch-backend plugin
 
 ## 0.1.3 - 2022-08-24